flask_travel_app/requirements.txt
Flask
python-dotenv
google-cloud-aiplatform
google-adk
gunicorn